requirements.txt
streamlit
pypdf
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.80
langchain-google-genai==2.1.12
langchain-text-splitters==0.3.11
langchain-huggingface==0.3.0
sentence-transformers
faiss-cpu